requirements.txt
fastapi
uvicorn
kafka-python
motor
odmantic
pydantic